refactor save methods
