+++ src/components/CreateGrants/NewGrantForm/helpers/mapArtifacts.ts
